Updated with corrections, fixed driver, renamed, generalized switchingfucnction file, move files around,  add in *.h Doxygen,
